+++ crates/polyfuse/src/lib.rs
@@ -1,9 +1,11 @@
 #![warn(clippy::todo, clippy::unimplemented)]
 
+mod decoder;
 mod session;
 mod util;
 mod write;
 
+pub mod bytes;
 pub mod op;
 pub mod reply;
